+++ src/Symfony/Component/Mailer/Bridge/Mailgun/CHANGELOG.md
@@ -1,7 +1,15 @@
 CHANGELOG
 =========
 
+4.4.0
+-----
+
+ * [BC BREAK] Renamed and moved `Symfony\Component\Mailer\Bridge\Mailgun\Http\Api\MailgunTransport`
+   to `Symfony\Component\Mailer\Bridge\Mailgun\Transport\MailgunApiTransport`, `Symfony\Component\Mailer\Bridge\Mailgun\Http\MailgunTransport`
+   to `Symfony\Component\Mailer\Bridge\Mailgun\Transport\MailgunHttpTransport`, `Symfony\Component\Mailer\Bridge\Mailgun\Smtp\MailgunTransport`
+   to `Symfony\Component\Mailer\Bridge\Mailgun\Transport\MailgunSmtpTransport`.
+
 4.3.0
 -----
 
- * added the bridge
+ * Added the bridge