+++ src/Symfony/Component/Mailer/Bridge/Postmark/CHANGELOG.md
@@ -1,7 +1,14 @@
 CHANGELOG
 =========
 
+4.4.0
+-----
+
+ * [BC BREAK] Renamed and moved `Symfony\Component\Mailer\Bridge\Postmark\Http\Api\PostmarkTransport`
+   to `Symfony\Component\Mailer\Bridge\Postmark\Transport\PostmarkApiTransport`, `Symfony\Component\Mailer\Bridge\Postmark\Smtp\PostmarkTransport`
+   to `Symfony\Component\Mailer\Bridge\Postmark\Transport\PostmarkSmtpTransport`.
+
 4.3.0
 -----
 
- * added the bridge
+ * Added the bridge